refactor: move ntrits hint to std/math/bits
