+++ backend/hint/builtin.go
@@ -19,16 +19,12 @@ var (
 
 	// NBits returns the n first bits of the input. Expects one argument: n.
 	NBits = NewStaticHint(nBits)
-
-	// NTrits returns the n first bits of the input. Expects one argument: n.
-	NTrits = NewStaticHint(nTrits)
 )
 
 func init() {
 	Register(IsZero)
 	Register(IthBit)
 	Register(NBits)
-	Register(NTrits)
 }
 
 func isZero(curveID ecc.ID, inputs []*big.Int, results []*big.Int) error {
@@ -68,21 +64,5 @@ func nBits(_ ecc.ID, inputs []*big.Int, results []*big.Int) error {
 	for i := 0; i < len(results); i++ {
 		results[i].SetUint64(uint64(n.Bit(i)))
 	}
-	return nil
-}
-
-func nTrits(_ ecc.ID, inputs []*big.Int, results []*big.Int) error {
-	n := inputs[0]
-	// TODO using big.Int Text method is likely not cheap
-	base3 := n.Text(3)
-	i := 0
-	for j := len(base3) - 1; j >= 0 && i < len(results); j-- {
-		results[i].SetUint64(uint64(base3[j] - 48))
-		i++
-	}
-	for ; i < len(results); i++ {
-		results[i].SetUint64(0)
-	}
-
 	return nil
 }
